+++ frontend/src/components/server/metaPanel/ChannelDetailPanel.tsx
@@ -9,8 +9,8 @@ import {
 } from 'semantic-ui-react'
 import UpdateChannelDescModal from 'components/server/modal/UpdateChannelDescModal'
 import UpdateChannelNameModal from 'components/server/modal/UpdateChannelNameModal'
-import { ChannelInfo } from '../redux/channel.slice'
-import { selectChannelUsers } from '../redux/channelUsers.slice'
+import { ChannelInfo } from 'components/server/redux/channels/channels.slice'
+import { selectChannelUsers } from 'components/server/redux/users/users.slice'
 
 interface ChannelDetailPanelProps {
     data: ChannelInfo
