+++ frontend/src/components/server/navbar/ServerNavbar.tsx
@@ -1,7 +1,7 @@
 import React, { FunctionComponent, useState } from 'react'
 import { useAppSelector } from 'redux/hooks'
 
-import { selectCurrentChannel } from 'components/server/redux/channel.slice'
+import { selectCurrentChannel } from 'components/server/redux/channels/channels.slice'
 
 import SearchModal from 'components/server/navbar/SearchModal'
 import ProfileDropdown from './ProfileDropdown'
@@ -18,7 +18,7 @@ const ServerNavbar: FunctionComponent<ServerNavbarProps> = () => {
             <div className="flex items-center justify-center relative">
                 <input
                     type="text"
-                    className="bg-slack-searchbar rounded-md w-1/2 text-slack-text-focus col-start-2 absolute placeholder-white px-4 cursor-pointer"
+                    className="bg-slack-searchbar rounded-md w-1/2 text-slack-text-focus col-start-2 fixed placeholder-white px-4 cursor-pointer"
                     placeholder={`Search something in #${currentChannel.name}`}
                     readOnly={true}
                     onClick={() => setIsSearchModalOpen(true)}
