+++ frontend/src/utils/channelUtil.ts
@@ -1,5 +1,5 @@
 import { UserInfo } from 'components/auth/redux/auth.slice'
-import { ChannelInfo } from 'components/server/redux/channel.slice'
+import { ChannelInfo } from 'components/server/redux/channels/channels.slice'
 import { CHANNEL_NAME_SEPARATOR } from '../constants/appConst'
 
 export const formatChannelName = (channelName: string) =>
